mv push/commit() helpers to tests/
