+++ tests/helpers/utils.py
@@ -4,6 +4,7 @@ import os
 import json
 import boto3
 import aurora_data_api
+import github
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
@@ -162,3 +163,41 @@ def check_ses_sender_email_auth(email_address: str, send_verify_email=False) -> 
             log.info("Sending SES verification email")
             ses.verify_email_identity(EmailAddress=email_address)
         return False
+
+
+def commit(repo, branch, changes, commit_message):
+    elements = []
+    for filepath, content in changes.items():
+        log.debug(f"Creating file: {filepath}")
+        blob = repo.create_git_blob(content, "utf-8")
+        elements.append(
+            github.InputGitTreeElement(
+                path=filepath, mode="100644", type="blob", sha=blob.sha
+            )
+        )
+
+    head_sha = repo.get_branch(branch).commit.sha
+    base_tree = repo.get_git_tree(sha=head_sha)
+    tree = repo.create_git_tree(elements, base_tree)
+    parent = repo.get_git_commit(sha=head_sha)
+    commit = repo.create_git_commit(commit_message, tree, [parent])
+
+    return commit
+
+
+def push(repo, branch, changes, commit_message="Adding test files"):
+    try:
+        ref = repo.get_branch(branch)
+    except Exception:
+        log.debug(f"Creating ref for branch: {branch}")
+        ref = repo.create_git_ref(
+            ref="refs/heads/" + branch,
+            sha=repo.get_branch(repo.default_branch).commit.sha,
+        )
+        log.debug(f"Ref: {ref.ref}")
+
+    commit_obj = commit(repo, branch, changes, commit_message)
+    log.debug(f"Pushing commit ID: {commit_obj.sha}")
+    ref.edit(sha=commit_obj.sha)
+
+    return branch
